Move USDZ export to the compiler
sdk/export.py was compiler machinery living in the SDK. Only
compiler/worker.py ever called export_object, and the agent-facing docs had
to tell the agent not to: "Do not call export_object from generated main.py".
A module the docs warn you away from is not part of the authoring API.

The line is declare vs derive. The SDK is what an object is declared and
checked with -- geometry, parts, articulations, materials, mass, TestContext.
Turning that declaration into a USDZ, with its UsdPhysics schemas, PBR
shading and audit, is what the compiler does with it.

The dependency direction improves too: this was sdk -> sdk, which hid that
export consumes the authoring API rather than belonging to it. It is now
compiler -> sdk.

_mass_solver and _collision stay in the SDK on purpose. testing.py and
visual.py need them for the agent's own checks, so they are shared kernels
rather than export machinery.
