+++ tests/test_collision_shapes.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 from pxr import Usd, UsdGeom, UsdPhysics
 
+from mini_articraft.compiler.export import export_object
 from mini_articraft.sdk import (
     ArticulatedObject,
     ArticulationType,
@@ -14,7 +15,6 @@ from mini_articraft.sdk import (
     Origin,
     SphereGeometry,
 )
-from mini_articraft.sdk.export import export_object
 from mini_articraft.sdk.mesh import boolean_difference
 
 # The usd-core stubs omit these schemas; bind them once instead of repeating the
