+++ tests/test_sdk_export.py
@@ -7,7 +7,8 @@ import pytest
 from build123d import Box, Pos
 from pxr import Gf, Usd, UsdGeom, UsdPhysics, UsdValidation
 
-import mini_articraft.sdk.export as export_module
+import mini_articraft.compiler.export as export_module
+from mini_articraft.compiler.export import export_object
 from mini_articraft.sdk import (
     ArticulatedObject,
     ArticulationType,
@@ -15,7 +16,6 @@ from mini_articraft.sdk import (
     MotionLimits,
     Origin,
 )
-from mini_articraft.sdk.export import export_object
 
 
 def test_export_writes_rigid_part_bodies_and_named_child_meshes(tmp_path) -> None:
