+++ tests/test_simulation_viewer.py
@@ -7,6 +7,7 @@ from pathlib import Path
 
 import pytest
 
+from mini_articraft.compiler.export import export_object
 from mini_articraft.sdk import (
     ArticulatedObject,
     ArticulationType,
@@ -15,7 +16,6 @@ from mini_articraft.sdk import (
     MotionLimits,
     Origin,
 )
-from mini_articraft.sdk.export import export_object
 from mini_articraft.viewer import load_viewer_run
 
 pytest.importorskip("mujoco", reason="recording motion needs the sim dependency group")
