+++ tests/test_texture_run.py
@@ -4,8 +4,8 @@ import json
 from types import SimpleNamespace
 
 import mini_articraft.compiler.worker as worker
+from mini_articraft.compiler.export import TextureExportReport, export_object
 from mini_articraft.sdk import ArticulatedObject, BoxGeometry
-from mini_articraft.sdk.export import TextureExportReport, export_object
 
 
 def _model() -> ArticulatedObject:
