+++ tests/test_viewer.py
@@ -11,8 +11,8 @@ import pytest
 from build123d import Box
 
 from mini_articraft import package_dir
+from mini_articraft.compiler.export import export_object
 from mini_articraft.sdk import ArticulatedObject, ArticulationType, MotionLimits, Origin
-from mini_articraft.sdk.export import export_object
 from mini_articraft.viewer import _handler, load_viewer_run
